+++ src/CleanRegex/Internal/Prepared/Template/Figure/MaskFigure.php
@@ -1,6 +1,7 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared\Template\Figure;
 
+use TRegx\CleanRegex\Internal\Flags;
 use TRegx\CleanRegex\Internal\Prepared\Phrase\Phrase;
 use TRegx\CleanRegex\Internal\Prepared\Template\Mask\KeywordsCondition;
 use TRegx\CleanRegex\Internal\Prepared\Template\Mask\MaskPhrase;
@@ -12,10 +13,10 @@ class MaskFigure implements Figure
     /** @var MaskPhrase */
     private $phrase;
 
-    public function __construct(string $mask, array $keywordsAndPatterns)
+    public function __construct(string $mask, Flags $flags, array $keywordsAndPatterns)
     {
         $this->condition = new KeywordsCondition($keywordsAndPatterns);
-        $this->phrase = new MaskPhrase($mask, $keywordsAndPatterns);
+        $this->phrase = new MaskPhrase($mask, $flags, $keywordsAndPatterns);
     }
 
     public function suitable(string $candidate): bool